+++ src/StandardTransitionStrategy.java
@@ -1,6 +1,6 @@
 public class StandardTransitionStrategy implements TransitionStrategy{
     @Override
-    public double getPrice(int weight) {
+    public double getPrice(float weight) {
         return 2.5 * weight;
     }
 }
